Got Add and edit  input for question working
