+++ src/components/section/modals/AddObj.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-const AddButton = styled.div.attrs({
+const Add = styled.div.attrs({
     className: 'btn btn-success',
 })`
     margin 15px 15px 15px 15px
@@ -9,9 +9,9 @@ const AddButton = styled.div.attrs({
 
 const AddObj = (props) => {
     return (
-        <AddButton onClick={props.onAddHandle}>
+        <Add onClick={props.onAddHandle}>
             Add {props.AddObjectName}
-        </AddButton>
+        </Add>
     )
 }
