+++ src/components/section/modals/Entry.js
@@ -1,21 +1,5 @@
 import React from 'react'
-import styled from 'styled-components'
-import { Options, ReadOnly, CustomOption, InputOption } from '../modals/EntryTypes'
-
-const Input = styled.input.attrs({
-    className: 'form-control',
-})`
-    margin: 5px;
-`
-
-const Label = styled.label`
-    margin: 1px;
-    font-size: 20px;
-`
-
-const Spacing = styled.div`
-    padding: 5ps;
-`
+import { Options, ReadOnly, CustomOption, InputOption, InputText } from '../modals/EntryTypes'
 
 const Entry = (props) => {        
     var list = ''
@@ -32,15 +16,12 @@ const Entry = (props) => {
 
     switch (type) {
         case 'InputText':
-            return (      
-                <Spacing>
-                    <Label>{props.label + ": "}</Label>       
-                    <Input
-                        type="text"
-                        value={props.value[props.label]}
-                        onChange={e => props.onChange(props.label, e.target.value)}
-                    />
-                </Spacing>  
+            return (   
+                <InputText
+                    label={props.label}
+                    value={props.value}
+                    onChange={props.onChange}
+                />
             )
         case 'Option':
             return (
@@ -88,18 +69,3 @@ const Entry = (props) => {
 }
 
 export default Entry
-
-// const InputText = (props) => {
-//     const { label, value, onChange} = props
-
-//     return (     
-//         <Spacing>
-//             <Label>{label + ": "}</Label>       
-//             <InputText
-//                 type="text"
-//                 value={value[label]}
-//                 onChange={e => onChange(props.label, e.target.value)}
-//             />
-//         </Spacing>  
-//     )
-// }
