+++ src/components/section/modals/EntryTypes/index.js
@@ -2,10 +2,12 @@ import Options from './Options'
 import ReadOnly from './ReadOnly'
 import CustomOption from './CustomOption'
 import InputOption from './InputOption'
+import InputText from './InputText'
 
 export {
     Options,
     ReadOnly,
     CustomOption,
-    InputOption
+    InputOption,
+    InputText
 }
